add two tests
A simple test to read dnstap input from a Unix domain socket and writes
nmsg to a UDP socket. This uses fstrm_replay to read existing testing
dnstap data and to send to the Unix domain socket and uses nmsgtool to
listen to the UDP socket for the nmsg data and writes in presentation
format.

The testing data was created using BIND9 to output to file.

A second simple test to read dnstap input over TCP and writes nmsg to
zeromq. This uses fstrm_replay to read existing testing dnstap data
and to send over TCP and uses nmsgtool to listen to the zeromq for
the nmsg data and writes in JSON format.